requirements.txt
pyqt5
pyqt5-sip
pyqtwebengine
pymupdf
qrcode
qtconsole
markdown
epc
retrying
python-magic
pygetwindow; sys_platform == 'win32'
